// src/openAPIGenerate/models/VnptPhatsoPrinterResponseModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type VnptPhatsoPrinterResponseModel = {
    id?: number;
    donviId?: number | null;
    name?: string | null;
    donviName?: string | null;
    enable?: boolean | null;
    passVerify?: string | null;
};