// src/openAPIGenerate/models/VnptPhatsoCallerResponseModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type VnptPhatsoCallerResponseModel = {
    id?: number;
    donviId?: number | null;
    callerName?: string | null;
    passVerify?: string | null;
    roomId?: number | null;
    roomName?: string | null;
    donviName?: string | null;
};