// src/openAPIGenerate/models/VnptPhatSoRoomGroupResponseModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type VnptPhatSoRoomGroupResponseModel = {
    id?: number;
    groupName?: string | null;
    donviId?: number | null;
    donviName?: string | null;
    enabled?: boolean | null;
    minNum?: number | null;
    maxNum?: number | null;
};