// src/openAPIGenerate/models/VnptPhatSoRoomResponseModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type VnptPhatSoRoomResponseModel = {
    id?: number;
    donviId?: number | null;
    roomName?: string | null;
    roomGroupId?: number | null;
    enabled?: boolean | null;
    lastPrintDate?: string | null;
    printedNum?: number | null;
    displayOrder?: number | null;
    startNum?: number | null;
    printName?: string | null;
    haveEmergency?: boolean | null;
    isBackup?: boolean | null;
    roomSubname?: string | null;
    roomGroupName?: string | null;
    donviName?: string | null;
};